3.11  - Mudei os tipos de variaveis de tipo double para  o tipo float. Coloquei uma função para colocar as restrições que forção a solução ser a que eu quero para a intância que da diferentes soluções para a heuristica.

+++ src/Modelo-Manne-2.0.cpp
@@ -27,10 +27,10 @@ int main(int argc, char **argv) {
 
 // Status da solução
 		int Status;
-		double SolucaoPrimal;
-		double SolucaoDual;
-		double Gap;
-		double Tempo;
+		float SolucaoPrimal;
+		float SolucaoDual;
+		float Gap;
+		float Tempo;
 		EscreveDadosLidosNaTela = 0;
